test_med/app/requirements.txt
django>=1.8
python-dateutil
more_itertools
nodeenv
django-bower
django-constance
